// src/types/types.ts
export interface Store {
  id: number;
  name: string;
  rating: number;
  reviews_count: number;
  location: string;
  overview: string;
  image: string;
  phone_number?: string;
  whats_app?: string;
  working_hours: string | null;
  is_verified: boolean;
  is_responsive: boolean;
  event_planning_categories: string[];
  rent_hire_categories: string[];
}

export interface StoreState {
  stores: Store[];
  selectedStore: Store | null;
  loading: boolean;
  error: string | null;
  store: Store | null;
}
// ...
export interface Offering {
  id: number; // ID as a string
  name: string;
  description: string;
  price: number | string;
  image: File | string | null; // Allow File, string, or null for fallback
  phoneNumber?: string;
  whatsappNumber?: string;
}
// ...
export interface StoreData {
  categories: string[]; // Store IDs as strings for consistency
  event_planning_categories: string[];
  rent_hire_categories: string[];
  name: string;
  phone_number: string;
  whats_app: string;
  image: File | null; // Use File | null for the form
  overview: string;
  location: string;
  working_hours: string;
  owner: null | string;
}